+++ components/brave_wallet_ui/options/auto_lock_options.ts
