+++ components/brave_wallet_ui/page/screens/backup-wallet/backup-recovery-phrase/backup-recovery-phrase.tsx
@@ -35,7 +35,7 @@ import {
 } from '../../onboarding/components/onboarding_content_layout/content_layout'
 import {
   SkipWarningDialog //
-} from '../explain-recovery-phrase/skip-warning-dialog'
+} from '../explain-recovery-phrase/skip_warning_dialog'
 
 // styles
 import {
