+++ components/brave_wallet_ui/page/screens/backup-wallet/explain-recovery-phrase/images/example_recovery_phrase_light.png
