+++ components/brave_wallet_ui/page/screens/backup-wallet/explain-recovery-phrase/skip_warning_dialog.tsx
