+++ components/brave_wallet_ui/page/screens/backup-wallet/verify-recovery-phrase/components/verification_progress.style.ts
