+++ components/brave_wallet_ui/page/screens/backup-wallet/verify-recovery-phrase/verify_recovery_phrase.style.ts
