+++ components/brave_wallet_ui/page/screens/backup-wallet/verify-recovery-phrase/verify_recovery_phrase.tsx
@@ -33,15 +33,15 @@ import Button from '@brave/leo/react/button'
 import {
   OnboardingContentLayout //
 } from '../../onboarding/components/onboarding_content_layout/content_layout'
-import { PhraseInput } from './components/phrase-input'
-import { VerificationProgress } from './components/verification-progress'
+import { PhraseInput } from './components/phrase_input'
+import { VerificationProgress } from './components/verification_progress'
 
 // styles
 import {
   Title //
 } from '../../onboarding/components/onboarding_content_layout/content_layout.style'
 import { Column, Row } from '../../../../components/shared/style'
-import { BackButton } from './verify-recovery-phrase.style'
+import { BackButton } from './verify_recovery_phrase.style'
 import { ContinueButton } from '../../onboarding/onboarding.style'
 
 const numberOfVerificationSteps = 3
