+++ components/brave_wallet_ui/page/screens/onboarding/components/onboarding_import_wallet_type/components/import_type_action.style.ts
