+++ components/brave_wallet_ui/page/screens/onboarding/components/onboarding_import_wallet_type/import_wallet_type.tsx
@@ -12,7 +12,7 @@ import { getLocale } from '../../../../../../common/locale'
 
 // components
 import { OnboardingContentLayout } from '../onboarding_content_layout/content_layout'
-import { ImportTypeAction } from './components/import-type-action'
+import { ImportTypeAction } from './components/import_type_action'
 
 // styles
 import { Column, VerticalDivider } from '../../../../../components/shared/style'
