+++ components/brave_wallet_ui/page/screens/onboarding/components/welcome_action/welcome_action.style.ts
