+++ components/brave_wallet_ui/page/screens/onboarding/connect-hardware/components/account_type.style.ts
