+++ components/brave_wallet_ui/page/screens/onboarding/connect-hardware/components/account_type.tsx
@@ -14,7 +14,7 @@ import {
   ArrowIcon,
   NetworkIcons,
   Description
-} from './account-type.style'
+} from './account_type.style'
 
 interface Props {
   title: string
