+++ components/brave_wallet_ui/page/screens/onboarding/create_password/components/create_password.tsx
