+++ components/brave_wallet_ui/page/screens/onboarding/create_password/onboarding_create_password.style.ts
